refactor: PlaceName VO 제거
